streaming: Use the optimal number of threads for VP8 encoding
We run the VP8 encoder in real time mode so it uses only the minimum
amount of time needed to encode each frame. However by default it
only uses one thread so that for large/complex frames it may run at
less than the source fps. Besides resulting in dropped frames this
blocks the main server thread for most of the time.
So this patch configures the VP8 encoder to use all the CPU's physical
core, resulting in less wall clock time spent in encode_frame().

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>