stream: implement interface for manual flush
The writing to network was always immediate.
Every write in the stream cause a write to the OS.
Can this have some penalty if you don't write large data as network
packets can be more fragmented or you encrypt data is smaller chunks
(when data are encrypted some padding is added then data is split in
multiple of encryption block which is usually the size of encryption
key and this is done for every write).

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>